Resolve review comments
