+++ gtsfm/loader/one_d_sfm_loader.py
@@ -17,9 +17,6 @@ from gtsfm.loader.loader_base import LoaderBase
 
 logger = logger_utils.get_logger()
 
-# Focal length is initialized to 1.2 * largest dimension of image if EXIF data is not available.
-NO_EXIF_DEFAULT_FOCAL_LENGTH_FACTOR = 1.2
-
 
 class OneDSFMLoader(LoaderBase):
     """Loader for datasets used in 1DSFM and Colmap papers.
@@ -41,7 +38,7 @@ class OneDSFMLoader(LoaderBase):
         folder: str,
         image_extension: str = "jpg",
         max_resolution: int = 640,
-        enable_no_exif: bool = False,
+        default_focal_length_factor: float = 0.0,
     ) -> None:
         """Initializes to load from a specified folder on disk.
 
@@ -52,15 +49,16 @@ class OneDSFMLoader(LoaderBase):
                 the smaller of the height/width of the image. e.g. for 1080p (1920 x 1080),
                 max_resolution would be 1080. If the image resolution max(height, width) is
                 greater than the max_resolution, it will be downsampled to match the max_resolution.
-            enable_no_exif: flag to whether to read images without exif.
+            default_focal_length_factor: focal length is initialized to default_focal_length_factor * largest dimension
+            of image if EXIF data is not available. Non-positive value means exif required.
         """
         super().__init__(max_resolution=max_resolution)
-        self._enable_no_exif = enable_no_exif
+        self._default_focal_length_factor = default_focal_length_factor
 
         # Fetch all the file names in /images folder.
         search_path = os.path.join(folder, "images", f"*.{image_extension}")
 
-        if not self._enable_no_exif:
+        if self._default_focal_length_factor > 0.0:
             self._image_paths = glob.glob(search_path)
         else:
             (self._image_paths, num_all_imgs) = self.get_images_with_exif(search_path)
@@ -69,27 +67,6 @@ class OneDSFMLoader(LoaderBase):
         self._num_imgs = len(self._image_paths)
         if self._num_imgs == 0:
             raise RuntimeError(f"Loader could not find any images with the specified file extension in {search_path}")
-
-    def get_images_with_exif(self, search_path: str) -> Tuple[List[str], int]:
-        """Return images with exif.
-        Args:
-            search_path: image sequence search path.
-        Returns:
-            Tuple[
-                List of image with exif paths.
-                The number of all the images.
-            ]
-        """
-        all_image_paths = glob.glob(search_path)
-        num_all_imgs = len(all_image_paths)
-        exif_image_paths = []
-        for single_img_path in all_image_paths:
-            # Drop images without exif.
-            if io_utils.load_image(single_img_path).get_intrinsics_from_exif() is None:
-                continue
-            exif_image_paths.append(single_img_path)
-
-        return (exif_image_paths, num_all_imgs)
 
     def image_filenames(self) -> List[str]:
         """Return the file names corresponding to each image index."""
@@ -128,16 +105,10 @@ class OneDSFMLoader(LoaderBase):
         Returns:
             Intrinsics for the given camera.
         """
-        # Get intrinsics from exif.
-        # TODO(yanwei) It is probably fine to only use the second method, as
-        # the images have been filtered before in consturctor.
-        if not self._enable_no_exif:
-            intrinsics = io_utils.load_image(self._image_paths[index]).get_intrinsics_from_exif()
-        else:
-            intrinsics = io_utils.load_image(self._image_paths[index]).get_intrinsics(
-                default_focal_length_factor=NO_EXIF_DEFAULT_FOCAL_LENGTH_FACTOR
-            )
-        return intrinsics
+        # Get intrinsics.
+        return io_utils.load_image(self._image_paths[index]).get_intrinsics(
+            default_focal_length_factor=self._default_focal_length_factor
+        )
 
     def get_camera_pose(self, index: int) -> Optional[Pose3]:
         """Get the camera pose (in world coordinates) at the given index.
